Pavios Chat Update 1
This is the first day of work and ill be doing more tomorrow. This app is going to get most of its ui from Chat GPT and iMessage. I want this experience to be seem less Basically I want the user to feel as though they are texting with the LLM. Their is a lot planed for this project so keep an eye out