Authentication
Almost done implementing Authentication. This will allow the user to connect to their own database or log in to my database. As of right now the data base will be used to store the chat history for your local models as well as if you are using the Open Ai API Key options, I will be using Email and password only right now as it is the easiest to setup for no tech users. Once authentication is complete I will be working on implementing the Open AI API request, then I will be working on adding the ability to communicate with local LLM models. I'm still looking into this but this is going to require probably Cloudflare tunnels but more info about that later